// src/interfaces/mantenedorprepolizas.model.ts
export class MantenedorPrePolizas {
  NumPoliza: string;
  NumCot: string;
  FecVigencia: string;
  CodTipoidenafi: number;
  NumIdenafi: string;
  NumCargas: number;
  NumOperacion: number;
  NumCorrelativo: number;
  CodTippension: string;
  CodEstcivil: string;
  CodAfp: string;
  CodIsapre: string;
  GlsDireccion: string;
  CodDireccion: number;
  GlsFono: string;
  GlsCorreo: string;
  CodViapago: string;
  CodSucursal: string;
  CodTipcuenta: string;
  CodBanco: string;
  NumCuenta: string;
  GlsNacionalidad: string;
  CodVejez: string;
  FecEmision: string;
  CodTipvia: string;
  GlsNomvia: string;
  GlsNumdmc: string;
  GlsIntdmc: string;
  CodTipzon: string;
  GlsNomzon: string;
  GlsReferencia: string;
  GlsFono2: string;
  GlsFono3: string;
  GlsCorreo2: string;
  GlsCorreo3: string;
  NumCuentacci: string;
  FecIngresosppText: string;
  GlsNomben: string;
  GlsNomsegben: string;
  GlsPatben: string;
  GlsMatben: string;
  CodSexo: string;
  FecNacben: string;
  FecFallben: string;
  FecInvben: string;
  CodCauinv: string;
  GlsDireccionCorresp: string;
  CodUbigeoCorresp: number;
  Departamento: string;
  CUSPP: string;
  NombreCompleto: string;
  Mensaje: string;
  DireccionStr: string;
  DirCorrStr: string;
  Usuario: string;
  FecAceptacion: string;
  Proteccion: string;
  Boleta: string;
}

export class PrePolizasBen {
  NumPoliza: string;
  NumOrden: number;
  CodGrufam: string;
  CodPar: string;
  CodSexo: string;
  CodStinv: string;
  CodCauinv: string;
  CodDerpen: string;
  CodDercre: string;
  CodTipoidenben: number;
  NumIdenben: string;
  GlsNomben: string;
  GlsNomsegben: string;
  GlsPatben: string;
  GlsMatben: string;
  FecNacben: string;
  FecInvben: string;
  FecFallben: string;
  FecNacHM: string;
  PrcPension: number;
  PrcPensionleg: number;
  MtoPension: number;
  MtoPensiongar: number;
  CodEstpension: string;
  PrcPensiongar: number;
  GlsFono: string;
  GlsFono2: string;
  GlsFono3: string;
  GlsCorreo: string;
  GlsCorreo2: string;
  GlsCorreo3: string;
  IndEstudiante: string;
  NumCot: string;
  NumOperacion: number;
  PrcPensionRep: number;
  ParentescoStr: string;
  TipoidenbenStr: string;
  Mensaje: string;
  Usuario: string;
  GlsDireccionBen: string;
  GlsDireccionCorrespBen: string;
  DireccionStrBen: string;
  DirCorrStrBen: string;
  CodUbigeoCorresp: number;
  CodDireccion: number;
  // Datos de los tutores
  MostrarTutor: string;
  NombreCompletoTut: string;
  PrimerNombreTut: string;
  SegundoNombreTut: string;
  ApellidoPaternoTut: string;
  ApellidoMaternoTut: string;
  TipoDocumentoTut: number;
  NumeroDocumentoTut: string;
  GlsDireccionTut: string;
  CodDireccionTut: number;
  GlsFono1Tut: string;
  GlsFono2Tut: string;
  GlsFono3Tut: string;
  GlsCorreo1Tut: string;
  GlsCorreo2Tut: string;
  CodViaPagoTut: string;
  CodSucursalTut: string;
  CodBancoTut: string;
  CodTipoCuentaTut: string;
  NumCuentaTut: string;
  NumCuentaCCITut: string;
  NumMesesPoder: number;
  FecVigenciaDesde: string;
  FecVigenciaHasta: string;
  FecEfecto: string;
  FecRecepcion: string;
  FechaTutorBen: string;
  ProteccionBen: string;
  BoletaBen: string;
}

export class ConsultaPrePolizas {
  TipoBusqueda: string;
  NumPoliza: string;
  TipoDocumento: number;
  NumDocumento: string;
  CUSPP: string;
  ApellidoPaterno: string;
  ApellidoMaterno: string;
  PrimerNombre: string;
  SegundoNombre: string;
}

export class CombosDireccion {
  GlsDistrito: string;
  CodProvinciaDis: string;
  CodDireccion: number;
  GlsProvincia: string;
  CodProvincia: string;
  CodRegionProv: string;
  GlsRegion: string;
  CodRegion: string;
}